Create Daily Dashboard with ratings information
This program daily extracts information from an internal database that contains information on ratings of television channels segmented by states in Mexico. For our television channel, we calculated the necessary metrics per tv program to understand the rating trends of each and their behavior against their respective competitors. The metrics calculated were yesterday's rating, the average rating of the last 7 days, and the average rating of the last 30 days and we calculate parameters such as rating in thousands, rating in percentage of households, average time viewed per person, carry over, among other metrics. We also calculate the same for competitors. Finally we calculate the rating of the last 30 programs and the rating grouped by month of the last 12 months.

Once calculating everything, we make a VBA macro that takes information from an excel sheet with the information we calculated and passes it to a dashboard in Excel where it already has all the cells formulated and the information is automatically updated with the ratings up to date.