Create Webscraping on real estate webpage.
This program detects when a query in forms has been filled with information about the characteristics of a real estate. It automatically detects the requested variables and does web scraping on a trusted page to extract all the results to an excel sheet. Finally it gives you a brief summary about the distribution of prices per square meter in the specified area.

When finishing the code you transform the ipynb as a exe to schedule the task on your PC.